+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
     <string name="title_activity_main">MainActivity</string>
 
     <string name="action_share">Share</string>
+    <string name="real_file">Realm file</string>
 </resources>
