+++ realm.io/build.gradle
@@ -22,4 +22,5 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile project(':benchmark')
     compile 'io.realm:realm-android:0.82.2'
+    compile 'com.github.dmytrodanylyk.realm-browser:library:0.0.3'
 }
